+++ SAMuel/My Project/AssemblyInfo.vb
@@ -10,8 +10,8 @@ Imports System.Runtime.InteropServices
 
 ' Review the values of the assembly attributes
 
-<Assembly: AssemblyTitle("SAMuel")> 
-<Assembly: AssemblyDescription("Service Application Manager Universal Expediting Library. SAMuel was designed and programmed by Associate CSR1, Zachary Karpinski, during the Summer of 2014. It automates or expediates various critical day to day tasks.")> 
+<Assembly: AssemblyTitle("SAMuel")>
+<Assembly: AssemblyDescription("SAMuel Automates Many Unambiguous Easy Labors. SAMuel was designed and programmed by Associate CSR1, Zachary Karpinski, during the Summer of 2014. It automates or expedites various critical day to day tasks.")>
 <Assembly: AssemblyCompany("Zachary Karpinski")> 
 <Assembly: AssemblyProduct("SAMuel")> 
 <Assembly: AssemblyCopyright("© 2014 Zachary Karpinski. All rights reserved.")> 
@@ -20,7 +20,7 @@ Imports System.Runtime.InteropServices
 <Assembly: ComVisible(False)> 
 
 'The following GUID is for the ID of the typelib if this project is exposed to COM
-<Assembly: Guid("d006cd8e-9b6d-4c33-a6f5-649c11ab716d")> 
+<Assembly: Guid("d006cd8e-9b6d-4c33-a6f5-649c11ab716d")>
 
 ' Version information for an assembly consists of the following four values:
 '
@@ -33,7 +33,6 @@ Imports System.Runtime.InteropServices
 ' by using the '*' as shown below:
 ' <Assembly: AssemblyVersion("1.0.*")> 
 
-<Assembly: AssemblyVersion("1.0.0.0")> 
-<Assembly: AssemblyFileVersion("1.1.0.0")> 
+<Assembly: AssemblyVersion("1.1.1.*")>
 
 <Assembly: NeutralResourcesLanguageAttribute("en-US")>