+++ SAMuel/frmOptions.Designer.vb
@@ -38,16 +38,19 @@ Partial Class FrmOptions
         Me.chkSuggestedAccountNumber = New System.Windows.Forms.CheckBox()
         Me.tabConversion = New System.Windows.Forms.TabPage()
         Me.btnDefaultPdfIni = New System.Windows.Forms.Button()
-        Me.TabControl1.SuspendLayout()
-        Me.tabGeneral.SuspendLayout()
-        Me.grpOptionsSaveFolders.SuspendLayout()
-        Me.tabOutlook.SuspendLayout()
-        Me.tabConversion.SuspendLayout()
-        Me.SuspendLayout()
+        Me.GroupBox1 = New System.Windows.Forms.GroupBox()
+        Me.btnSetupPdfIni = New System.Windows.Forms.Button()
+        Me.TabControl1.SuspendLayout
+        Me.tabGeneral.SuspendLayout
+        Me.grpOptionsSaveFolders.SuspendLayout
+        Me.tabOutlook.SuspendLayout
+        Me.tabConversion.SuspendLayout
+        Me.GroupBox1.SuspendLayout
+        Me.SuspendLayout
         '
         'txtSavePath
         '
-        Me.txtSavePath.Enabled = False
+        Me.txtSavePath.Enabled = false
         Me.txtSavePath.Location = New System.Drawing.Point(62, 8)
         Me.txtSavePath.Name = "txtSavePath"
         Me.txtSavePath.Size = New System.Drawing.Size(346, 20)
@@ -60,7 +63,7 @@ Partial Class FrmOptions
         Me.btnSelectSaveFolder.Size = New System.Drawing.Size(75, 23)
         Me.btnSelectSaveFolder.TabIndex = 1
         Me.btnSelectSaveFolder.Text = "Save To.."
-        Me.btnSelectSaveFolder.UseVisualStyleBackColor = True
+        Me.btnSelectSaveFolder.UseVisualStyleBackColor = true
         '
         'btnSaveOptions
         '
@@ -69,11 +72,11 @@ Partial Class FrmOptions
         Me.btnSaveOptions.Size = New System.Drawing.Size(75, 23)
         Me.btnSaveOptions.TabIndex = 2
         Me.btnSaveOptions.Text = "Save"
-        Me.btnSaveOptions.UseVisualStyleBackColor = True
+        Me.btnSaveOptions.UseVisualStyleBackColor = true
         '
         'lblSaveTo
         '
-        Me.lblSaveTo.AutoSize = True
+        Me.lblSaveTo.AutoSize = true
         Me.lblSaveTo.Location = New System.Drawing.Point(2, 11)
         Me.lblSaveTo.Name = "lblSaveTo"
         Me.lblSaveTo.Size = New System.Drawing.Size(60, 13)
@@ -83,8 +86,8 @@ Partial Class FrmOptions
         'TabControl1
         '
         Me.TabControl1.Controls.Add(Me.tabGeneral)
-        Me.TabControl1.Controls.Add(Me.tabOutlook)
         Me.TabControl1.Controls.Add(Me.tabConversion)
+        Me.TabControl1.Controls.Add(Me.tabOutlook)
         Me.TabControl1.Location = New System.Drawing.Point(1, 1)
         Me.TabControl1.Name = "TabControl1"
         Me.TabControl1.SelectedIndex = 0
@@ -104,7 +107,7 @@ Partial Class FrmOptions
         Me.tabGeneral.Size = New System.Drawing.Size(498, 166)
         Me.tabGeneral.TabIndex = 0
         Me.tabGeneral.Text = "General"
-        Me.tabGeneral.UseVisualStyleBackColor = True
+        Me.tabGeneral.UseVisualStyleBackColor = true
         '
         'btnViewLog
         '
@@ -113,7 +116,7 @@ Partial Class FrmOptions
         Me.btnViewLog.Size = New System.Drawing.Size(75, 23)
         Me.btnViewLog.TabIndex = 5
         Me.btnViewLog.Text = "View Logfile"
-        Me.btnViewLog.UseVisualStyleBackColor = True
+        Me.btnViewLog.UseVisualStyleBackColor = true
         '
         'grpOptionsSaveFolders
         '
@@ -123,7 +126,7 @@ Partial Class FrmOptions
         Me.grpOptionsSaveFolders.Name = "grpOptionsSaveFolders"
         Me.grpOptionsSaveFolders.Size = New System.Drawing.Size(91, 69)
         Me.grpOptionsSaveFolders.TabIndex = 4
-        Me.grpOptionsSaveFolders.TabStop = False
+        Me.grpOptionsSaveFolders.TabStop = false
         Me.grpOptionsSaveFolders.Text = "Save Folders"
         '
         'btnShow
@@ -133,7 +136,7 @@ Partial Class FrmOptions
         Me.btnShow.Size = New System.Drawing.Size(75, 23)
         Me.btnShow.TabIndex = 1
         Me.btnShow.Text = "Show"
-        Me.btnShow.UseVisualStyleBackColor = True
+        Me.btnShow.UseVisualStyleBackColor = true
         '
         'btnEmpty
         '
@@ -142,7 +145,7 @@ Partial Class FrmOptions
         Me.btnEmpty.Size = New System.Drawing.Size(75, 23)
         Me.btnEmpty.TabIndex = 0
         Me.btnEmpty.Text = "Empty"
-        Me.btnEmpty.UseVisualStyleBackColor = True
+        Me.btnEmpty.UseVisualStyleBackColor = true
         '
         'tabOutlook
         '
@@ -154,67 +157,88 @@ Partial Class FrmOptions
         Me.tabOutlook.Size = New System.Drawing.Size(498, 166)
         Me.tabOutlook.TabIndex = 1
         Me.tabOutlook.Text = "Outlook"
-        Me.tabOutlook.UseVisualStyleBackColor = True
+        Me.tabOutlook.UseVisualStyleBackColor = true
         '
         'chkAuditMode
         '
-        Me.chkAuditMode.AutoSize = True
+        Me.chkAuditMode.AutoSize = true
         Me.chkAuditMode.Location = New System.Drawing.Point(7, 9)
         Me.chkAuditMode.Name = "chkAuditMode"
         Me.chkAuditMode.Size = New System.Drawing.Size(107, 17)
         Me.chkAuditMode.TabIndex = 15
         Me.chkAuditMode.Text = "Audit each email."
-        Me.chkAuditMode.UseVisualStyleBackColor = True
+        Me.chkAuditMode.UseVisualStyleBackColor = true
         '
         'chkSuggestedAccountNumber
         '
-        Me.chkSuggestedAccountNumber.AutoSize = True
+        Me.chkSuggestedAccountNumber.AutoSize = true
         Me.chkSuggestedAccountNumber.Location = New System.Drawing.Point(7, 32)
         Me.chkSuggestedAccountNumber.Name = "chkSuggestedAccountNumber"
         Me.chkSuggestedAccountNumber.Size = New System.Drawing.Size(182, 17)
         Me.chkSuggestedAccountNumber.TabIndex = 5
         Me.chkSuggestedAccountNumber.Text = "Use suggested account numbers"
-        Me.chkSuggestedAccountNumber.UseVisualStyleBackColor = True
+        Me.chkSuggestedAccountNumber.UseVisualStyleBackColor = true
         '
         'tabConversion
         '
-        Me.tabConversion.Controls.Add(Me.btnDefaultPdfIni)
+        Me.tabConversion.Controls.Add(Me.GroupBox1)
         Me.tabConversion.Location = New System.Drawing.Point(4, 22)
         Me.tabConversion.Name = "tabConversion"
         Me.tabConversion.Padding = New System.Windows.Forms.Padding(3)
         Me.tabConversion.Size = New System.Drawing.Size(498, 166)
         Me.tabConversion.TabIndex = 2
         Me.tabConversion.Text = "Conversion"
-        Me.tabConversion.UseVisualStyleBackColor = True
+        Me.tabConversion.UseVisualStyleBackColor = true
         '
         'btnDefaultPdfIni
         '
-        Me.btnDefaultPdfIni.Location = New System.Drawing.Point(7, 137)
+        Me.btnDefaultPdfIni.Location = New System.Drawing.Point(18, 19)
         Me.btnDefaultPdfIni.Name = "btnDefaultPdfIni"
         Me.btnDefaultPdfIni.Size = New System.Drawing.Size(103, 23)
         Me.btnDefaultPdfIni.TabIndex = 1
-        Me.btnDefaultPdfIni.Text = "Default PDF995"
-        Me.btnDefaultPdfIni.UseVisualStyleBackColor = True
+        Me.btnDefaultPdfIni.Text = "Restore Default"
+        Me.btnDefaultPdfIni.UseVisualStyleBackColor = true
+        '
+        'GroupBox1
+        '
+        Me.GroupBox1.Controls.Add(Me.btnSetupPdfIni)
+        Me.GroupBox1.Controls.Add(Me.btnDefaultPdfIni)
+        Me.GroupBox1.Location = New System.Drawing.Point(3, 3)
+        Me.GroupBox1.Name = "GroupBox1"
+        Me.GroupBox1.Size = New System.Drawing.Size(147, 160)
+        Me.GroupBox1.TabIndex = 2
+        Me.GroupBox1.TabStop = false
+        Me.GroupBox1.Text = "PDF995 Printer Settings"
+        '
+        'btnSetupPdfIni
+        '
+        Me.btnSetupPdfIni.Location = New System.Drawing.Point(17, 52)
+        Me.btnSetupPdfIni.Name = "btnSetupPdfIni"
+        Me.btnSetupPdfIni.Size = New System.Drawing.Size(103, 23)
+        Me.btnSetupPdfIni.TabIndex = 2
+        Me.btnSetupPdfIni.Text = "Setup Modified"
+        Me.btnSetupPdfIni.UseVisualStyleBackColor = true
         '
         'FrmOptions
         '
-        Me.AutoScaleDimensions = New System.Drawing.SizeF(6.0!, 13.0!)
+        Me.AutoScaleDimensions = New System.Drawing.SizeF(6!, 13!)
         Me.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font
         Me.ClientSize = New System.Drawing.Size(506, 227)
         Me.Controls.Add(Me.TabControl1)
         Me.Controls.Add(Me.btnSaveOptions)
         Me.Name = "FrmOptions"
         Me.Text = "Options"
-        Me.TabControl1.ResumeLayout(False)
-        Me.tabGeneral.ResumeLayout(False)
-        Me.tabGeneral.PerformLayout()
-        Me.grpOptionsSaveFolders.ResumeLayout(False)
-        Me.tabOutlook.ResumeLayout(False)
-        Me.tabOutlook.PerformLayout()
-        Me.tabConversion.ResumeLayout(False)
-        Me.ResumeLayout(False)
+        Me.TabControl1.ResumeLayout(false)
+        Me.tabGeneral.ResumeLayout(false)
+        Me.tabGeneral.PerformLayout
+        Me.grpOptionsSaveFolders.ResumeLayout(false)
+        Me.tabOutlook.ResumeLayout(false)
+        Me.tabOutlook.PerformLayout
+        Me.tabConversion.ResumeLayout(false)
+        Me.GroupBox1.ResumeLayout(false)
+        Me.ResumeLayout(false)
 
-    End Sub
+End Sub
     Friend WithEvents txtSavePath As System.Windows.Forms.TextBox
     Friend WithEvents FolderBrowserDialog1 As System.Windows.Forms.FolderBrowserDialog
     Friend WithEvents btnSelectSaveFolder As System.Windows.Forms.Button
@@ -231,4 +255,6 @@ Partial Class FrmOptions
     Friend WithEvents chkAuditMode As System.Windows.Forms.CheckBox
     Friend WithEvents tabConversion As System.Windows.Forms.TabPage
     Friend WithEvents btnDefaultPdfIni As System.Windows.Forms.Button
+    Friend WithEvents GroupBox1 As GroupBox
+    Friend WithEvents btnSetupPdfIni As Button
 End Class