+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 2015-01-20
+
+Features:
+
+  - -
+
+Code Improvements:
+
+  - AddContact: Script improved.
+  - TDrive: Fields cleaned up and changed formatting.
+
 ## 2014-12-14
 
 Features: